feat: adjust downloading to handle multiple files
