+++ renderer/ions/atoms/index.ts
@@ -22,3 +22,4 @@ export const imagesAtom = atom<
 export const directoryAtom = atom("");
 export const modelDownloadNoteAtom = atomWithStorage("modelDownloadNote", true);
 export const captionRunningAtom = atomWithStorage("captionRunning", false);
+export const checkpointsAtom = atom<string[]>([]);
