+++ renderer/public/locales/de/common.json
@@ -17,7 +17,7 @@
   "delete": "Löschen",
   "directory": "Verzeichnis",
   "download": "Herunterladen",
-  "downloading": "wird heruntergeladen",
+  "downloading": "Wird heruntergeladen",
   "enterKeyToUseGPTVision": " Geben Sie Ihren OpenAI-API-Schlüssel ein, um GPT-Vision zu verwenden. Russian (RU)",
   "exampleResponse": "Beispielantwort",
   "feedback": "Feedback",
