+++ renderer/public/locales/en/common.json
@@ -17,7 +17,7 @@
   "delete": "Delete",
   "directory": "Directory",
   "download": "Download",
-  "downloading": "downloading",
+  "downloading": "Downloading",
   "enterKeyToUseGPTVision": "Enter your OpenAI API key to use GPT-Vision.",
   "exampleResponse": "Example Response",
   "feedback": "Feedback",
