+++ renderer/public/locales/es/common.json
@@ -17,7 +17,7 @@
   "delete": "Eliminar",
   "directory": "Directorio",
   "download": "Descargar",
-  "downloading": "descargando",
+  "downloading": "Descargando",
   "enterKeyToUseGPTVision": "Introduce tu clave de API de OpenAI para usar GPT-Vision.",
   "exampleResponse": "Respuesta de Ejemplo",
   "feedback": "Comentarios",
