+++ renderer/public/locales/fr/common.json
@@ -17,7 +17,7 @@
   "delete": "Supprimer",
   "directory": "Répertoire",
   "download": "Télécharger",
-  "downloading": "téléchargement",
+  "downloading": "Téléchargement",
   "enterKeyToUseGPTVision": "Entrez votre clé API OpenAI pour utiliser GPT-Vision.",
   "exampleResponse": "Exemple de Réponse",
   "feedback": "Retours",
