+++ renderer/public/locales/it/common.json
@@ -17,7 +17,7 @@
   "delete": "Elimina",
   "directory": "Directory",
   "download": "Scarica",
-  "downloading": "scaricando",
+  "downloading": "Scaricando",
   "enterKeyToUseGPTVision": "Inserisci la tua chiave API OpenAI per utilizzare GPT-Vision.",
   "exampleResponse": "Esempio di Risposta",
   "feedback": "Feedback",
