+++ renderer/public/locales/nl/common.json
@@ -17,7 +17,7 @@
   "delete": "Verwijderen",
   "directory": "Map",
   "download": "Downloaden",
-  "downloading": "downloaden",
+  "downloading": "Downloaden",
   "enterKeyToUseGPTVision": "Voer uw OpenAI API-sleutel in om GPT-Vision te gebruiken.",
   "exampleResponse": "Voorbeeldantwoord",
   "feedback": "Feedback",
