+++ renderer/public/locales/pl/common.json
@@ -17,7 +17,7 @@
   "delete": "Usuń",
   "directory": "Katalog",
   "download": "Pobierz",
-  "downloading": "pobieranie",
+  "downloading": "Pobieranie",
   "enterKeyToUseGPTVision": "Wprowadź swój klucz API OpenAI, aby korzystać z GPT-Vision.",
   "exampleResponse": "Przykładowa Odpowiedź",
   "feedback": "Opinie",
