+++ renderer/public/locales/pt/common.json
@@ -17,7 +17,7 @@
   "delete": "Excluir",
   "directory": "Diretório",
   "download": "Baixar",
-  "downloading": "baixando",
+  "downloading": "Baixando",
   "enterKeyToUseGPTVision": "Insira sua chave de API da OpenAI para usar o GPT-Vision.",
   "exampleResponse": "Exemplo de Resposta",
   "feedback": "Feedback",
